engine_traits: rename cf.rs to cfdefs.rs
